avancée quizz crud and entity

+++ challenge-devops-api/src/main/java/com/esiea/challengedevopsapi/entities/User.java
@@ -12,7 +12,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     private String login;
     private String password;
@@ -32,13 +32,11 @@ public class User {
     public User() {
     }
 
-    // --- Getters and setters ---
-
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
